Rearrange .auth content into separated files
